refactor(user-settings): reorganize profile settings components and update imports
