+++ src/app/(profile)/profile/settings/page.tsx
@@ -1,4 +1,4 @@
-import { ProfileSettingsScreen } from '@frontend/features/setting-page/profile-settings-screen';
+import { ProfileSettingsScreen } from '@frontend/features/user-settings/profile-settings-screen';
 
 export default function SettingsPage() {
   return <ProfileSettingsScreen />;
